states: Ensure error with NBD_REPLY_TYPE_ERROR_*
A server is non-compliant if it replies with an NBD_REPLY_TYPE_ERROR_*
chunk but no error value.  Rather than accidentally treat such a
command as passing, let's convert things to EINVAL instead.  (If we
later add a strictness knob, this is yet another thing to gate when
the user wants to validate a server's compliance).