uri: Reject URIs not in scheme:// form
libxml2 parses nbd:unix:/path/to/socket as valid per RFC 3986, using
the nbd: scheme for TCP, no authority (thus defaulting to
localhost:10809), and a relative path 'unix:/path/to/socket' as the
export name. But this string has historically been used with qemu to
request a Unix socket with no export name, such that nbdkit --run
produces it for $nbd (compared to $unixsocket with just the socket
name, or the new $uri in the next nbdkit release), which can be quite
confusing when you end up with a failed connection or worse silently
connecting to a completely different server than intended.

As RFC 3986 permits particular schemes to require the use of only a
subset of what the overall URI grammar will permit, it is easiest to
just require that all use of nbd: or nbd+unix: schemes as URIs always
use the form with an authority field (even when the authority is the
empty string), which in turn implies the path element will either be
empty or begin with '/'.

Reported-by: Martin Kletzander <mkletzan@redhat.com>