socket: Avoid SIGPIPE where possible
If the server hangs up while we are writing, we'd rather have send()
fail with EPIPE than deal with a SIGPIPE signal - especially when we
are being run as a library inside a larger program that may have other
intentions on what signals should do for its other activities.  POSIX
requires the MSG_NOSIGNAL flag which makes this easy (it requires that
we use send() instead of write(), which in turn requires socketpair()
rather than pipe() for nbd_connect_command - but we already do that).
If we ever branch out into accepting pre-opened fds, we'll have to
deal with SIGPIPE suppression for non-sockets at that time.