poll: Improve our interface
Make nbd_poll slightly more like poll(), allowing a user to detect
timeouts by returning 0 on timeout and 1 when we made progress. It
turns out that none of our internal uses ever expect a timeout (we
only call nbd_internal_poll with timeout==-1 because we expect a reply
from the server), but the public function might as well be nicer.

Also handle POLLERR (server closed its read end, so our writes will
fail - this is a typical real-life problem) and POLLNVAL (the fd
itself is closed - most likely a bug in our program) as outright
errors, and POLLHUP the same as POLLIN (the server closed its write
end, so read() will eventually see EOF even if we have to drain a
buffer first).

Perhaps we also want to add an nbd_aio_notify_err() that can inform
the machine of any externally-detected error on the fd to the server,
where the poll loop invokes that on POLLERR, and where we tweak the
generator to permit that external event in any other public state.
But that's a bigger patch.