generator: Add support for namespace constants
This just defines the namespace, its contexts and related constants and the only
supported one is currently base:allocation.  The names of the constants are not
very future-proof, but shorter than LIBNBD_META_NS_CONTEXT_BASE_ALLOCATION or
similar.

Currently the output looks like this:

/* "base" namespace */

/* "base" namespace contexts */

/* "base:allocation" context related constants */

Separated by two empty lines from unrelated parts of the header file above and
below.

Signed-off-by: Martin Kletzander <mkletzan@redhat.com>