maint: Use $(NULL) for all Makefile.am macro lists
This borrows from a trick in libvirt - by defining $(NULL) to expand
to an empty string, we can more consistently write multi-line macros
where all useful lines terminate with \, making it easier to
add/remove lines without worrying about whether \ needs to be touched
up on neighboring lines.

Mainly a mechanical search for '= \' coupled with a manual fix to the
later lines of each affected macro.