tests: Make errors more robust under load
When run under valgrind, the 'errors' test would fail about 80% of the
time (on my setup) because a single NBD_CMD_WRITE managed to actually
send() the entire packet to the server without blocking the state
machine.  To make things more robust, switch to a server which is
serialized (memory is parallel, but sh is serial), and which
intentionally does not read a second command until the first is
processed, then switch the client to send two commands rather than one
(whether or not the first loses the race with blocking, we can now
guarantee the second will block).

Since we are now writing a shell script, it is also easy to use that
script to add another improvement: check that a server error
propagates back to the client as expected. Surprisingly, this
uncovered an nbdkit bug (now fixed) - it was very difficult to coerce
nbdkit sh into failing with anything other than EIO, because it
mistakenly used strcmp() for output ending in a trailing space but not
a trailing newline.