api: Swap callback arguments in pread_structured
In the pread_structured callback, it was odd that we placed an in/out
argument (int *error) in the middle of two input-only arguments.  It's
more obvious now that we also have the block_status and *_notify
callbacks all taking the error argument last.  Swap the parameter
order for consistency.  This is an API break, but we have not declared
stable API yet; and callers can use LIBNBD_HAVE_AIO_PREAD_NOTIFY as a
witness of another API change made in the same release if they must
support earlier versions as well.