docs: s/handle/cookie/ when referring to nbd_aio_FOO result
It is confusing to read about 'handle' meaning both the struct
nbd_handle encompassing the entire state machine (a typical name for
the opaque object used to interact with a library), and a command
handle being the 63-bit identifier returned from nbd_aio_FOO to later
retire in nbd_aio_command_completed (because it gets stored in the
field named 'handle' in the NBD spec).  Change the latter usage to
'cookie' where it makes sense (chosen in preference to 'id').  No
API/ABI change results, but it does cause a bit of churn.