generator: Define new Closure type instead of callbacks.
A Closure is a list of (usually one, but can be more) closures.  In C
there is also a singe ‘void *user_data’ parameter which is passed by
the caller into the function and through as the first parameter of
each callback invocation.

By grouping the previously separate Opaque and Callback* parameters
together we can avoid the awkward situation where we have to scan
through the argument list to try to match them up.

Because this is a closure, in non-C languages we can simply map these
to a closure and drop the opaque parameter entirely.  It is not needed
since languages with proper closures can capture local state in the
closure.

Unlike the previous code it is no longer possible to mix persistent
and non-persistent callbacks in the same API call.  This was not used
before and is unlikely to be useful.

For the C API there is no API or ABI change (the only change is the
naming of the opaque pointer which is not part of the API).  For the
non-C languages the opaque parameter is no longer required as
discussed above.

Partly based on Eric Blake's earlier work here:
https://www.redhat.com/archives/libguestfs/2019-July/thread.html#00160