api: Allow completion callbacks to auto-retire
When using the nbd_aio_FOO_callback commands, there is nothing further
to be learned about the command by calling nbd_aio_command_completed()
compared to what the callback already had access to.  There are still
scenarios where manually retiring the command after the fact may be
useful (whether the return was 0 to keep the status unchanged, or -1
to alter the retirement status to *error), but by allowing a return
value of 1, we can reduce the number of API calls required.  We can
also auto-retire the lone NBD_CMD_DISC request, reducing the amount of
special casing in nbd_aio_[peek_]command_completed.

Note that although this is not an API change, it does change semantics
(any client which happened to return 1 but still retire commands will
now see different behavior); but that's okay as we have not declared a
stable interface yet.

Update a couple of the tests and examples to give this coverage:
examples/glib-main-loop no longer piles up unretired commands, and
tests/server-death shows that even a command stranded by server death
can still be auto-retired.

The docs are also updated to mention the rare race where it may be
possible to see a server's reply without blocking such that the
callback fires before the original API completes (we may change this
later by deferring the callback if we detect it was triggered by the
current API), and to reduce the amount of copy-and-paste in the API
reference by instead referring to a common section in the main
documentation.