tests: Fix server-death under load
Under valgrind, server-death was slowed down enough that nbkdit was
able to queue up the reply to NBD_CMD_TRIM on the wire prior to the
test kill()ing the server, and the nbd_aio_poll then read the response
to the trim rather than the desired action of seeing the command be
stranded by server death.  Fix it by adding a delay so that the server
does not respond to trim in a timely manner.  While at it, adjust the
read delay to be longer (just in case of a heavily loaded system), and
use the preferred spelling of the --filter=delay parameters.

Fixes:7e53d48e
Reported-by: Richard W.M. Jones <rjones@redhat.com>