lib: Call read/extent(FREE) before callback(VALID|FREE)
Some callers of nbd_aio_pread_structured_callback or
nbd_aio_block_status_callback share the same opaque pointer to both
calls, with the intent to only free the pointer on the completion
callback.  Although our documentation does not explicitly specify the
order in which callbacks are made, allowing the callback(FREE) to
occur prior to the read(FREE) puts a burden on the client to not
dereference the user_data in the read() function if the VALID bit is
not set.  If the client follows our suggested practice of an immediate
return 0 if the VALID bit is not set, then this is the case, but we
can't guarantee all clients will use that pattern.  And, it's fairly
easy to guarantee that ALL of our read/extent callbacks are made prior
to any of our callback() use, by utilizing the FINISH state of
structured replies to catch the few cases where we are unable to use
read/extent(VALID|FREE) because the server used a separate
NBD_REPLY_TYPE_NONE instead of packing the NBD_REPLY_FLAG_DONE bit to
an earlier chunk.

Note that nbd_internal_retire_and_free_command still has to check
whether we have freed the callbacks, as nbd_close() can strand
commands before the FINISH state.