lib: Use symbol versions.
This changes the libnbd.syms file (the linker script) to use symbol
versions.  The only version currently used is ‘LIBNBD_1.0’ for a
notional stable 1.0 release in the future.  It is expected that future
stable releases will use ‘LIBNBD_1.2’ etc.

Note this only deals with new symbols, not with the question of how to
deal with symbols which we decide to change incompatibly, if that
issue ever arises.