examples: Fix theoretical cookie race in example.
There was a theoretic race in this example: If the server was very
fast at handling commands then it's possible that in a call such as:

  cookie = nbd_aio_pread_callback (..., callback, ...);
  buffers[i].cookie = cookie;

nbd_aio_pread_callback completes the command and calls the callback
before returning.  buffers[i].cookie would therefore not be set before
the callback() function checks the buffer list for the cookie.  This
would have caused an abort() in the existing code, although we have
never observed that.

The new code, instead of relying on the cookie, passes a pointer to
&buffers[i], so the callback no longer needs to search the list of
buffers (this is also quicker).

This necessitated another change however: previously we were copying
buffers around to ensure that the next free buffer was always at
&buffers[nr_buffers].  Instead of that introduce a new state
(BUFFER_UNUSED) and search the short list of buffers once when looking
for a free buffer.

Thanks: Eric Blake.