tests: Make sure errors test uses serialized nbdkit
An upcoming nbdkit release will be relaxing sh scripts to be fully
parallel unless the script requests otherwise; but our errors test
depends on the current serialized execution.  Thankfully, older nbdkit
does not care if we provide the new hook for the sake of newer nbdkit.

An audit for other testsuite use of nbdkit sh turns up
eflags-plugin.sh and meta-base-allocation.sh, but those will continue
to work in a parallel model.