ocaml: Make OCaml callbacks return int instead of unit.
Callbacks in C have always returned an int, but this was previously
mapped to 0 (OK) or -1 (error), and therefore from the OCaml bindings
we could just assume a ‘unit’ return type (like void in C), with
exceptions mapped to errors.

However commit 7e53d48ede71e66f changed the behaviour of the
completion callback to allow 1 to mean auto-retire.  To map this to
OCaml we must change the return type of callbacks from ‘unit’ to
‘int’.

Note that error is still indicated by throwing an exception.