ocaml: Add NBD.Buffer alloc and free.
NBD.Buffer.alloc is a convenience function to allocate an
uninitialized buffer of a particular size (instead of having to do
‘NBD.Buffer.of_bytes (Bytes.create size)’).

NBD.Buffer.free manually frees the C buffer.  This is necessary when
writing certain types of AIO code so you can free the buffer in the
completion callback.  In fact it's not really necessary since you can
pass the buffer to the completion callback and just do nothing with
it: the buffer will then go out of scope at the right time so the
destructor will free the buffer.  But this is awkward to get right so
having an explicit free function makes it more obvious what is
happening.