lib: Print function call parameters and results in debug output.
This adds function call parameters and results when debugging is
enabled.  It looks like this:

  libnbd: debug: nbd1: nbd_connect_unix: enter: unixsocket="/tmp/socket"
  libnbd: debug: nbd1: nbd_connect_unix: leave: ret=0

When changing the handle name you see:

  libnbd: debug: nbd1: nbd_create: opening handle
  libnbd: debug: nbd1: nbd_set_handle_name: enter: handle_name="src"
  libnbd: debug: src: nbd_set_handle_name: leave: ret=0

The output is enabled whenever debugging is on, however because of
limitations in the way debugging works it can only be done on
functions which have may_set_error = true.

This prints full strings, but omits lists of strings and buffers.  We
probably in future want it to truncate strings, escape non-printable
characters, print parts of buffers and so on.

In libguestfs we had different flags for debugging and tracing, but
one lesson we learned from that is that we almost always want both to
be enabled and often have to go back to bug reporters to get them to
reproduce bugs with both enabled.