generator: Rename RConstString -> RStaticString and improve documentation.
RConstString sounded like it could apply to any ‘const char *’.
However this is not the case.  For example if nbd_get_handle_name
returned the h->hname field as an RConstString then there would be a
race if another thread updated the handle name.

To avoid confusion, rename this to RStaticString, and in the
documentation make it clear that these strings are located in the text
section and survive until the process exits (or libnbd is dlclosed if
using dynamic loading).

Thanks: Eric Blake