ocaml/tests: Fix race when creating and deleting files in tests.
Two of the OCaml tests used a test file in the local directory with a
fixed name.  However because each test runs twice (once for bytecode
and once for native code), on a machine with many cores the file might
be deleted by one instance of the test while it is still in use by the
other instance, causing a test failure.  (This was observed on a
machine with 48 cores.)