api: In public documentation describe %CLOSED state as "shut down".
In libnbd-api(3):

  "The handle must be connected and finished handshaking with the
   server, or closed, otherwise this call will return an error."

You can call functions such as nbd_can_flush when the handle is in the
%CLOSED state.  This state doesn't mean that the handle has been
closed (by nbd_close), but is more like a clean shutdown (nbd_shutdown
or receiving EOF from the server).  In particular, if the handle
reaches the %CLOSED state then it must have completed the handshake at
some point earlier so these server flags will be valid.

To avoid confusion, the public documentation now refers to this state
as "shut down".  The updated documentation reads:

  "The handle must be connected and finished handshaking with the
   server, or shut down, otherwise this call will return an error."