generator: Change handling of Flags to be a true optional argument.
In libguestfs generator we have the concept of optargs which is
different from plain arguments.  These are mapped to optional
arguments in languages that support them such as Python.

This commit adds a new concept of optargs.  At the moment it is simply
limited to handling the optional (in some bindings) flags parameter
which is used to handle the NBD_CMD_FLAG_* flags.

If present, the old Flags parameter becomes OFlags and is moved into
the optargs list.

For the OCaml generation this change simplifies things considerably as
we no longer need the mapping from C arg to ocaml_arg (they are now
the same).

In the libguestfs C bindings the handling of optargs is rather
complex, and I don't intend to replicate that here.  Instead they are
just handled as non-optional arguments appearing after the normal
arguments.

Note this commit does not change the API in any language.