generator: Handle closure args (cbargs) specially.
If we accept that callbacks will never handle the full range of
parameters then we can simplify the generator quite a bit by using a
special type for closure args vs normal method args.

This removes many asserts and quite a bit of unreachable code
(eg. Python code for handling ArrayAndLen in normal methods that was
never used).

The output of the generator after this commit should be identical.

It's possible to go a little further if we wanted: CBArrayAndLen is
only ever used for ‘uint32_t’ arrays.  CBMutable is only ever used for
‘int*’.  We could make CB* types which only handle those cases.