generator: Generalize OFlags.
In a future commit we want to add (non-optional) Flags arg.  As a step
to doing this, generalize OFlags so it's not tied to just
NBD_CMD_FLAG_*.

This does not change the C API.

It does introduce a small change to the OCaml API -- putting related
flags into a submodule, basically renaming them.  Note we don't
provide guarantees for non-C APIs.