ocaml/tests: Compact heap after every test to detect memory corruption.
After each test add a call to Gc.compact.  This will iterate over the
entire heap and compact the memory.  A side effect of this is it
causes every pointer to be followed and all unreferenced objects to be
finalized, which is a simple way to detect if there's any memory
corruption in the OCaml bindings.