generator: Add Enum type for enumerated types / unions.
Previously nbd_set_tls had an integer argument which was 0 for
disable, 1 for allow and 2 for require.  This commit adds a proper
enumerated type to describe this, defining LIBNBD_TLS_DISABLE = 0,
LIBNBD_TLS_ALLOW = 1 and LIBNBD_TLS_REQUIRE = 2.  (Note the C ABI
doesn't change).

In C the enumerated type is still defined and passed as an int (not as
an enum).  While we could define an enum type for this, there are ABI
stability problems inherent in enums.

In OCaml this is implemented as a variant type.

There is no equivalent for returning an enum (eg. for nbd_get_tls).
We should add that later.  It won't affect the C API but would change
the OCaml API.