ocaml: Add a standard test for doing asynch copy from one handle to another.
For non-C languages add a standard test where we do an asynch copy
from one nbdkit server to another via two libnbd handles.  The first
nbdkit server is serving standard data read-only using
nbdkit-pattern-plugin.  The second instance is running a RAM disk to
receive the data.  The copy is done in a loop which issues a certain
number of simultaneous asynchronous in-flight reads and then uses
completion callbacks to issue writes.

This tests all kinds of features of libnbd such as asynchronous I/O,
multiple handles in the same process, persistent buffers and
completion callbacks.

The OCaml version of the test is copied from one of the examples,
which a compatible license change.

This does not yet implement a Python version of this test since trying
to create one has revealed a number of bugs in the Python bindings.