generator: Share single list of all Closures.
This change does not affect the output of the generator.

Note this requires that all Closure args have the same parameter name
whichever method they appear in.  An alternate refactoring could work
the same way as Flags and Enum where the parameter name is part of the
arg, eg:

  type arg =
  ...
  | Closure of string * closure   (* name, type *)
  ...
  "set_debug_callback", {
    default_call with
    args = [ Closure ("debug", debug_closure) ];

So after this change, Closure, Flags, and Enum aren't quite congruent.