api: Add nbd_clear_debug_callback.
Commit 0904dd113dfa36485623b3c1756dc1aeab3dddeb removed the
theoretical possibility of clearing the debug callback from the handle
by setting it to NULL (although that was dubious from an API point of
view).

This commit adds an explicit way to do this.  Another advantage of
this is we can internally call this API from other places when we want
to clear/free the debug callback.