api: Combine callback and user_data into a single struct.
The definition of functions that take a callback is changed so that
the callback and user_data are combined into a single structure, eg:

  int64_t nbd_aio_pread (struct nbd_handle *h,
            void *buf, size_t count, uint64_t offset,
-           int (*completion_callback) (/*..*/), void *user_data,
+           nbd_completion_callback completion_callback,
            uint32_t flags);

Several nbd_*_callback structures are defined.  The one corresponding
to the example above is:

  typedef struct {
    int (*callback) (unsigned valid_flag, void *user_data, int *error);
    void *user_data;
  } nbd_completion_callback;

The nbd_aio_pread function can now be called using:

  nbd_aio_pread (nbd, buf, sizeof buf, offset,
                 (nbd_completion_callback) { .callback = my_fn,
                                             .user_data = my_data },
                 0);

Note that the whole structure is passed by value not by reference.

For OClosure only, a NULL callback can be passed using this macro:

  nbd_aio_pread (nbd, buf, sizeof buf, offset, NBD_NULL_COMPLETION, 0);