docs: Replace libnbd-api(3) with man pages for individual APIs.
These are generated with the exception of nbd_create(3) which is hand
written; and nbd_close(3), nbd_get_error(3) and nbd_get_errno(3) which
are links to nbd_create(3).

Note that unusually two generated files are checked into git,
docs/.gitignore and docs/Makefile.inc.  That is to make it easier to
autoreconf straight from a fresh git clone.