build: Drop docs/.gitignore
.gitignore allows negative match patterns, useful for specifying a
glob to ignore then exceptions that matched the glob to not ignore
after all.  With judicious use of negative patterns for the few files
we maintain by hand, we can do everything needed for ignoring built
files in docs/ without having to generate and commit a docs/.gitignore
(reverting part of commit 65165ca7).