maint: We require GNU make and C99
The build fails without GNU make because we've used things like
$(wildcard), and it's easy enough to get gmake on BSD machines that we
aren't worried about rewriting things to be portable to POSIX make.

Likewise, the build fails with a C89 compiler; C99 is easily available
enough now (we may later decide to require __attribute__((cleanup))
like nbdkit, which narrows the compilers to modern gcc or clang, but
for now the untested assumption is that another C99 compiler will
still work).